Bike Computer - Diagnostics Version
Use this file during testing and development of the platform. The compiled size is considerably larger than the file without Serial, so don't use it if you don't need to.
Once running, simply connect to serial at 9600 bps and you will get useful info.
Info will be displayed as key = value, where key is an input or output, and value is the value of the key. All info will be on the same line, seperated by a tab. Each line is the next tick